requirements.txt
streamlit
streamlit-option-menu
google-generativeai
easyocr
pillow
numpy
requests
beautifulsoup4
regex
librosa
soundfile
scikit-learn
joblib
